Update to XXH3 for CRC_OPT

+++ src/PaletteTexture.h
@@ -15,7 +15,7 @@ public:
 private:
 	CachedTexture * m_pTexture;
 	u8* m_pbuf;
-	u32 m_paletteCRC256;
+	u64 m_paletteCRC256;
 };
 
 extern PaletteTexture g_paletteTexture;
